hand feature product

+++ app/Services/Attribute/Value/AttributeValueService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services\Attribute\Value;
 
+use App\Http\Resources\Attribute\Value\ValueResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Http\Traits\Paginate;
 use App\Repositories\Attribute\AttributeRepositoryInterface;
@@ -29,28 +30,34 @@ class AttributeValueService  implements AttributeValueServiceInterface
         $sort = request()->query('sort') ?? 'desc';
         if($sort !== 'desc' && $sort !== 'asc') $sort = 'asc';
         $values = $attribute->values()->orderBy($column, $sort);
-        return $this->loadRelationships($values)->get();
+        return ValueResource::collection($this->loadRelationships($values)->get());
     }
     public function create(int|string $aid,array $data){
         $attribute = $this->attributeRepository->find($aid);
         if(!$attribute) throw new ModelNotFoundException('Attribute not found');
         $value = $attribute->values()->create($data);
-        return $this->loadRelationships($value);
+        return new ValueResource($this->loadRelationships($value));
     }
     public function find(int|string $aid,int|string $id){
         $attribute = $this->attributeRepository->find($aid);
         if(!$attribute) throw new ModelNotFoundException('Attribute not found');
         $value = $attribute->values()->find($id);
         if(!$value) throw new ModelNotFoundException('Attribute value not found');
-        return $value;
+        return new ValueResource($this->loadRelationships($value));
     }
     public function update(int|string $aid,int|string $id, array $data){
-        $value = $this->find($aid,$id);
+        $attribute = $this->attributeRepository->find($aid);
+        if(!$attribute) throw new ModelNotFoundException('Attribute not found');
+        $value = $attribute->values()->find($id);
+        if(!$value) throw new ModelNotFoundException('Attribute value not found');
         $value->update($data);
-        return $this->loadRelationships($value);
+        return new ValueResource($this->loadRelationships($value));
     }
     public function delete(int|string $aid,int|string $id){
-        $value = $this->find($aid,$id);
+        $attribute = $this->attributeRepository->find($aid);
+        if(!$attribute) throw new ModelNotFoundException('Attribute not found');
+        $value = $attribute->values()->find($id);
+        if(!$value) throw new ModelNotFoundException('Attribute value not found');
         $value->delete();
         return true;
     }
